khadidja_obesityproject/requirements.txt
absl-py==1.4.0
altair==5.0.1
annotated-types==0.5.0
anyio==3.6.2
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
astunparse==1.6.3
attrs==23.1.0
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.0.0
blinker==1.6.2
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
chatbot==1.5.2b0
ChatterBot==1.0.4
chatterbot-corpus==1.2.0
click==8.1.3
cmdstanpy==1.1.0
colorama==0.4.6
comm==0.1.3
contourpy==1.0.7
convertdate==2.4.0
cycler==0.11.0
Cython==3.0.0
dacite==1.8.1
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.6
docopt==0.6.2
ephem==4.1.4
et-xmlfile==1.1.0
executing==1.2.0
fastjsonschema==2.17.1
feedparser==6.0.10
filelock==3.12.2
Flask==2.2.3
flatbuffers==23.3.3
fonttools==4.38.0
fqdn==1.5.1
frozendict==2.3.8
gast==0.4.0
gitdb==4.0.10
GitPython==3.1.31
google-auth==2.16.3
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.51.3
h5py==3.8.0
holidays==0.28
html5lib==1.1
htmlmin==0.1.12
idna==3.4
ImageHash==4.3.1
imageio==2.26.0
importlib-metadata==6.7.0
importlib-resources==6.0.0
ipykernel==6.23.1
ipython==8.13.2
ipython-genutils==0.2.0
ipywidgets==8.0.6
isoduration==20.11.0
itsdangerous==2.1.2
jax==0.4.6
jedi==0.18.2
Jinja2==3.1.2
joblib==1.1.1
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.6.3
jupyter_client==8.2.0
jupyter_core==5.3.0
jupyter_server==2.6.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.7
kaggle==1.5.16
keras==2.13.1
kiwisolver==1.4.4
libclang==15.0.6.1
LunarCalendar==0.0.9
lxml==4.9.2
Markdown==3.4.3
markdown-it-py==3.0.0
MarkupSafe==2.1.2
mathparse==0.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mdurl==0.1.2
mglearn==0.1.9
missingno==0.5.2
mistune==2.0.5
mpmath==1.2.1
multimethod==1.9.1
multitasking==0.0.11
nbclassic==1.0.0
nbclient==0.8.0
nbconvert==7.4.0
nbformat==5.8.0
nest-asyncio==1.5.6
networkx==3.0rc1
nltk==3.8.1
notebook==6.5.4
notebook_shim==0.2.3
numpy==1.23.5
oauthlib==3.2.2
opencv-contrib-python==4.8.0.74
opencv-python==4.8.0.76
openpyxl==3.1.2
opt-einsum==3.3.0
overrides==7.3.1
packaging==23.0
pandas==1.5.3
pandas-datareader==0.10.0
pandas-profiling==3.2.0
pandocfilters==1.5.0
parso==0.8.3
patsy==0.5.3
phik==0.12.3
pickleshare==0.7.5
Pillow==9.3.0
Pint==0.20.1
pipenv==2023.6.26
pipreqs==0.4.13
platformdirs==3.5.1
plot==0.6.5
plotly==5.15.0
pmdarima==2.0.3
prometheus-client==0.17.0
prompt-toolkit==3.0.38
prophet==1.1.4
protobuf==4.22.0
psutil==5.9.5
pure-eval==0.2.2
pyarrow==12.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.10.12
pydantic-settings==2.0.3
pydantic_core==2.6.0
pydeck==0.8.1b0
Pygments==2.15.1
PyMeeus==0.5.12
pymongo==3.13.0
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.19.3
pytesseract==0.3.10
python-dateutil==2.8.2
python-dotenv==1.0.0
python-json-logger==2.0.7
python-slugify==8.0.1
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.4.1
pywin32==306
pywinpty==2.0.10
PyYAML==6.0
pyzmq==25.0.2
qtconsole==5.4.3
QtPy==2.3.1
regex==2022.10.31
requests==2.28.1
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.4.2
rsa==4.9
scikit-learn==1.2.1
scipy==1.10.1
seaborn==0.12.2
Send2Trash==1.8.2
sgmllib3k==1.0.0
six==1.16.0
sklearn==0.0.post1
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.4.1
SQLAlchemy==1.2.19
stack-data==0.6.2
statsmodels==0.14.0
streamlit==1.24.0
streamlit-option-menu==0.3.6
streamlit-pandas-profiling==0.1.3
sympy==1.11.1
tabulate==0.9.0
tangled-up-in-unicode==0.2.0
tenacity==8.2.2
tensorboard==2.13.0
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorflow==2.12.0
tensorflow-estimator==2.13.0
tensorflow-intel==2.13.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.2.0
terminado==0.17.1
text-unidecode==1.3
tflearn==0.5.0
threadpoolctl==3.1.0
tinycss2==1.2.1
toml==0.10.2
toolz==0.12.0
torchaudio==2.1.0.dev20230324+cu118
torchvision==0.16.0.dev20230324+cu118
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
typeguard==2.13.3
typing==3.7.4.3
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==4.3.1
uri-template==1.2.0
urllib3==1.26.13
validators==0.20.0
virtualenv==20.23.1
virtualenv-clone==0.5.7
visions==0.7.4
watchdog==3.0.0
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.5.2
Werkzeug==2.2.3
widgetsnbextension==4.0.7
windows-curses==2.3.1
wordcloud==1.9.2
wrapt==1.14.1
xlrd==2.0.1
yarg==0.1.9
ydata-profiling==4.5.1
yfinance==0.2.22
zipp==3.15.0
